create a java class to modify the appearance of Transaction elements
