+++ app/src/main/java/com/example/expensetracker/Transaction.java
@@ -68,4 +68,3 @@ public class Transaction {
         this.note = description;
     }
 }
-
